+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# NEXT RELEASE
+
 ## 0.9.0
 * Apply extracted text to non-Tweets, non-Reddit posts.
 * `--delete-originals` option