+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT RELEASE
 
+### 0.9.2
+
 ## 0.9.0
 * Apply extracted text to non-Tweets, non-Reddit posts.
 * `--delete-originals` option